refactor(docs): button dx
