+++ apps/docs/content/components/button/custom-impl.raw.jsx
@@ -1,0 +1,38 @@
+import {forwardRef} from "react";
+import {useButton, Ripple, Spinner} from "@nextui-org/react";
+
+const MyButton = forwardRef((props, ref) => {
+  const {
+    domRef,
+    children,
+    spinnerSize,
+    spinner = <Spinner color="current" size={spinnerSize} />,
+    spinnerPlacement,
+    startContent,
+    endContent,
+    isLoading,
+    disableRipple,
+    getButtonProps,
+    getRippleProps,
+  } = useButton({
+    ref,
+    ...props,
+  });
+
+  const {ripples, onClear} = getRippleProps();
+
+  return (
+    <button ref={domRef} {...getButtonProps()}>
+      {startContent}
+      {isLoading && spinnerPlacement === "start" && spinner}
+      {children}
+      {isLoading && spinnerPlacement === "end" && spinner}
+      {endContent}
+      {!disableRipple && <Ripple ripples={ripples} onClear={onClear} />}
+    </button>
+  );
+});
+
+MyButton.displayName = "MyButton";
+
+export default MyButton;
