+++ apps/docs/content/components/button/custom-styles.raw.jsx
@@ -1,0 +1,12 @@
+import {Button} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Button
+      className="bg-gradient-to-tr from-pink-500 to-yellow-500 text-white shadow-lg"
+      radius="full"
+    >
+      Button
+    </Button>
+  );
+}
