+++ apps/docs/content/components/button/disabled.raw.jsx
@@ -1,0 +1,9 @@
+import {Button} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Button isDisabled color="primary">
+      Button
+    </Button>
+  );
+}
