+++ apps/docs/content/components/button/group-disabled.raw.jsx
@@ -1,0 +1,11 @@
+import {Button, ButtonGroup} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <ButtonGroup isDisabled>
+      <Button>One</Button>
+      <Button>Two</Button>
+      <Button>Three</Button>
+    </ButtonGroup>
+  );
+}
