+++ apps/docs/content/components/button/group-disabled.ts
@@ -1,14 +1,4 @@
-const App = `import {Button, ButtonGroup} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <ButtonGroup isDisabled>
-      <Button>One</Button>
-      <Button>Two</Button>
-      <Button>Three</Button>
-    </ButtonGroup>
-  );
-}`;
+import App from "./group-disabled.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
