+++ apps/docs/content/components/button/loading.raw.jsx
@@ -1,0 +1,9 @@
+import {Button} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Button isLoading color="primary">
+      Loading
+    </Button>
+  );
+}
