+++ apps/docs/content/components/button/loading.ts
@@ -1,12 +1,4 @@
-const App = `import {Button} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Button color="primary" isLoading>
-      Loading
-    </Button>
-  );
-}`;
+import App from "./loading.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
