+++ apps/docs/content/components/button/radius.raw.jsx
@@ -1,0 +1,13 @@
+import {Button} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex gap-4 items-center">
+      <Button radius="full">Full</Button>
+      <Button radius="lg">Large</Button>
+      <Button radius="md">Medium</Button>
+      <Button radius="sm">Small</Button>
+      <Button radius="none">None</Button>
+    </div>
+  );
+}
