+++ apps/docs/content/components/button/radius.ts
@@ -1,26 +1,4 @@
-const App = `import {Button} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4 items-center">
-      <Button radius="full">
-        Full
-      </Button>
-      <Button radius="lg">
-        Large
-      </Button>  
-      <Button radius="md">
-        Medium
-      </Button>  
-      <Button radius="sm">
-        Small
-      </Button>  
-      <Button radius="none">
-        None
-      </Button>  
-    </div>
-  );
-}`;
+import App from "./radius.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
