+++ apps/docs/content/components/button/sizes.raw.jsx
@@ -1,0 +1,11 @@
+import {Button} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex gap-4 items-center">
+      <Button size="sm">Small</Button>
+      <Button size="md">Medium</Button>
+      <Button size="lg">Large</Button>
+    </div>
+  );
+}
