+++ apps/docs/content/components/button/usage.raw.jsx
@@ -1,0 +1,5 @@
+import {Button} from "@nextui-org/react";
+
+export default function App() {
+  return <Button color="primary">Button</Button>;
+}
